+++ src/loaders/zarrLoader.js
@@ -1,6 +1,6 @@
 import { BoundsCheckError } from 'zarr';
 
-import { guessRgb, padTileWithZeros, byteSwapInplace } from './utils';
+import { guessRgb, byteSwapInplace } from './utils';
 import { DTYPE_VALUES } from '../constants';
 import HTTPStore from './httpStore';
 
@@ -59,10 +59,12 @@ export default class ZarrLoader {
 
   /**
    * Returns image tiles at tile-position (x, y) at pyramidal level z.
-   * @param {number} x positive integer
-   * @param {number} y positive integer
-   * @param {number} z positive integer (0 === highest zoom level)
-   * @param {Array} loaderSelection, Array of valid dimension selections
+   * @param {number} args
+   * @param {number} args.x positive integer
+   * @param {number} args.y positive integer
+   * @param {number} args.z positive integer (0 === highest zoom level)
+   * @param {Array} args.loaderSelection Array of valid dimension selections
+   * @param {Array} args.signal AbortSignal object
    * @returns {Object} data: TypedArray[], width: number (tileSize), height: number (tileSize)
    */
   async getTile({ x, y, z, loaderSelection = [], signal }) {
@@ -94,30 +96,20 @@ export default class ZarrLoader {
         // big endian
         byteSwapInplace(data);
       }
-      const width = source.chunks[xIndex];
-      const height = source.chunks[yIndex];
-      if (height < this.tileSize || width < this.tileSize) {
-        return padTileWithZeros(
-          { data, width, height },
-          this.tileSize,
-          this.tileSize
-        );
-      }
       return data;
     });
     const data = await Promise.all(dataRequests);
     if (source.store instanceof HTTPStore && signal?.aborted) return null;
-    return {
-      data,
-      width: this.tileSize,
-      height: this.tileSize
-    };
+    const width = source.chunks[xIndex];
+    const height = source.chunks[yIndex];
+    return { data, width, height };
   }
 
   /**
    * Returns full image panes (at level z if pyramid)
-   * @param {number} z positive integer (0 === highest zoom level)
-   * @param {Array} loaderSelection, Array of valid dimension selections
+   * @param {number} args
+   * @param {number} args.z positive integer (0 === highest zoom level)
+   * @param {Array} args.loaderSelection, Array of valid dimension selections
    * @returns {Object} data: TypedArray[], width: number, height: number
    */
   async getRaster({ z, loaderSelection = [] }) {
@@ -187,7 +179,8 @@ export default class ZarrLoader {
 
   /**
    * Returns image width and height (at pyramid level z) without fetching data
-   * @param {number} z positive integer (0 === highest zoom level)
+   * @param {Object} args
+   * @param {number} args.z positive integer (0 === highest zoom level)
    * @returns {Object} width: number, height: number
    */
   getRasterSize({ z }) {