+++ src/viewers/SideBySideViewer.js
@@ -1,6 +1,6 @@
 import React from 'react'; // eslint-disable-line import/no-unresolved
 import VivViewer from './VivViewer';
-import { SideBySideView } from '../views';
+import { SideBySideView, getDefaultInitialViewState } from '../views';
 
 /**
  * This component provides a side-by-side VivViewer with linked zoom/pan.
@@ -9,14 +9,18 @@ import { SideBySideView } from '../views';
  * @param {Array} props.colorValues List of [r, g, b] values for each channel.
  * @param {Array} props.channelIsOn List of boolean values for each channel for whether or not it is visible.
  * @param {string} props.colormap String indicating a colormap (default: '').  The full list of options is here: https://github.com/glslify/glsl-colormap#glsl-colormap
- * @param {Object} props.loader Loader to be used for fetching data.  It must have the properies `dtype`, `numLevels`, and `tileSize` and implement `getTile` and `getRaster`.
+ * @param {Object} props.loader Loader to be used for fetching data.  It must have the properies `dtype`, `numLevels`, `isPyramid`, and `tileSize` and implement `getTile`, `getRaster`, and `getRasterSize`.
  * @param {Array} props.loaderSelection Selection to be used for fetching data.
  * @param {Boolean} props.zoomLock Whether or not lock the zooms of the two views.
  * @param {Boolean} props.panLock Whether or not lock the pans of the two views.
- * @param {boolean} props.isLensOn Whether or not to use the lens.
- * @param {number} props.lensSelection Numeric index of the channel to be focused on by the lens.
- * @param {number} props.lensBorderColor RGB color of the border of the lens (default [255, 255, 255]).
- * @param {number} props.lensBorderRadius Percentage of the radius of the lens for a border (default 0.02).
+ * @param {Object} props.initialViewState Object like { target: [x, y, 0], zoom: -zoom } for initializing where the viewer looks (optional - this can be inferred from height/width/loader).
+ * @param {number} props.height Current height of the component.
+ * @param {number} props.width Current width of the component.
+ * @param {boolean} [props.isLensOn] Whether or not to use the lens deafult (false).
+ * @param {number} [props.lensSelection] Numeric index of the channel to be focused on by the lens (default 0).
+ * @param {Array} [props.lensBorderColor] RGB color of the border of the lens (default [255, 255, 255]).
+ * @param {number} [props.lensBorderRadius] Percentage of the radius of the lens for a border (default 0.02).
+ * @param {import('./VivViewer').ViewStateChange} [props.onViewStateChange] Callback that returns the deck.gl view state (https://deck.gl/docs/api-reference/core/deck#onviewstatechange).
  */
 const SideBySideViewer = props => {
   const {
@@ -29,24 +33,33 @@ const SideBySideViewer = props => {
     panLock,
     loaderSelection,
     zoomLock,
+    height,
+    width,
     isLensOn = false,
-    lensSelection,
+    lensSelection = 0,
     lensRadius = 100,
     lensBorderColor = [255, 255, 255],
-    lensBorderRadius = 0.02
+    lensBorderRadius = 0.02,
+    onViewStateChange
   } = props;
+  const viewState =
+    initialViewState || getDefaultInitialViewState(loader, { height, width });
   const detailViewLeft = new SideBySideView({
-    initialViewState: { ...initialViewState, id: 'left' },
+    initialViewState: { ...viewState, id: 'left' },
     linkedIds: ['right'],
     panLock,
-    zoomLock
+    zoomLock,
+    height,
+    width: width / 2
   });
   const detailViewRight = new SideBySideView({
-    initialViewState: { ...initialViewState, id: 'right' },
-    x: initialViewState.width,
+    initialViewState: { ...viewState, id: 'right' },
+    x: width / 2,
     linkedIds: ['left'],
     panLock,
-    zoomLock
+    zoomLock,
+    height,
+    width: width / 2
   });
   const layerConfig = {
     loader,
@@ -64,7 +77,12 @@ const SideBySideViewer = props => {
   const views = [detailViewRight, detailViewLeft];
   const layerProps = [layerConfig, layerConfig];
   return loader ? (
-    <VivViewer layerProps={layerProps} views={views} randomize />
+    <VivViewer
+      layerProps={layerProps}
+      views={views}
+      randomize
+      onViewStateChange={onViewStateChange}
+    />
   ) : null;
 };