+++ src/views/DetailView.js
@@ -1,14 +1,18 @@
 import { MultiscaleImageLayer, ImageLayer, ScaleBarLayer } from '../layers';
 import VivView from './VivView';
 import { getVivId } from './utils';
+import { OVERVIEW_VIEW_ID } from './OverviewView';
+
+export const DETAIL_VIEW_ID = 'detail';
 
 /**
  * This class generates a MultiscaleImageLayer and a view for use in the VivViewer as a detailed view.
+ * It takes the same arguments for its constructor as its base class VivView.
  * */
 export default class DetailView extends VivView {
   getLayers({ props, viewStates }) {
     const { loader } = props;
-    const { id } = this;
+    const { id, height, width } = this;
     const layerViewState = viewStates[id];
     const layers = [];
 
@@ -34,12 +38,28 @@ export default class DetailView extends VivView {
             loader,
             unit,
             size: value,
-            viewState: layerViewState
+            viewState: { ...layerViewState, height, width }
           })
         );
       }
     }
 
     return layers;
+  }
+
+  /**
+   * Create a viewState for this class, checking the id to make sure this class and viewState match.
+   * @param {Object} args
+   * @param {ViewState} args.ViewState ViewState object.
+   * @returns {ViewState} ViewState for this class (or null by default if the ids do not match).
+   */
+  filterViewState({ viewState, currentViewState }) {
+    if (viewState.id === OVERVIEW_VIEW_ID) {
+      const { target } = viewState;
+      if (target) {
+        return { ...currentViewState, target };
+      }
+    }
+    return super.filterViewState({ viewState });
   }
 }