+++ src/views/index.js
@@ -1,5 +1,6 @@
-export { default as DetailView } from './DetailView';
-export { default as OverviewView } from './OverviewView';
+export { default as DetailView, DETAIL_VIEW_ID } from './DetailView';
+export { default as OverviewView, OVERVIEW_VIEW_ID } from './OverviewView';
 export { default as VivView } from './VivView';
 export { default as SideBySideView } from './SideBySideView';
 export { default as Static3DView } from './Static3DView';
+export { getDefaultInitialViewState } from './utils';